fastapi/requirements.txt
anyio==3.6.2
attrs==22.2.0
boto==2.49.0
boto3==1.26.74
botocore==1.29.74
certifi==2022.12.7
click==8.1.3
exceptiongroup==1.1.0
fastapi==0.92.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
packaging==23.0
pluggy==1.0.0
pydantic==1.10.5
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
rfc3986==1.5.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
